use opts instead of config for barbar, add global merge_tables function
