+++ init.lua
@@ -1,3 +1,4 @@
+require('globals')
 require('init-package-manager')
 require('plugins')
 require('options')
